backend/apps/challenges/serializers.py
from rest_framework import serializers
from .models import Challenge


class ChallengeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Challenge
        fields = ["id", "name", "start_date", "end_date", "created_at"]