backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
drf-spectacular==0.27.2
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0